+++ hutool-core/src/main/java/cn/hutool/core/io/checksum/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * IO校验相关库和工具
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.core.io.checksum;